Revert "Started replacing sprite indices (`uint16_t`) by a class `Sprite`"
This reverts commit 5071234cab9055804cc1c41ff85629574a8be82c.

Reverting to leave the project in a working state in case it is left on its own
for a while.

+++ engine/libflatkiss-data/libflatkiss/data/loader_animation_player.hpp
@@ -33,7 +33,7 @@ class LoaderAnimationPlayer {
       std::string const& file_path);
 
  private:
-  static std::unordered_map<int64_t, Animation> loadGroup(
+  static std::unordered_map<uint16_t, Animation> loadGroup(
       int64_t group_size, std::ifstream& animations_stream);
 };
